requirements.txt
fastapi
uvicorn
pyyaml
openai
pydantic
